requirements.txt
blurhash==1.1.4
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==5.1.1
idna==3.10
Mastodon.py==1.8.1
ndg-httpsclient==0.5.1
python-dateutil==2.8.2
python-magic==0.4.27
requests==2.32.3
six==1.16.0
urllib3==2.2.3
